🔨Refactor: 계좌 금액 충전
별개의 js 파일로 수정
